Restructure login page
