+++ ui/package.json
@@ -23,6 +23,7 @@
     "react-dom": "^18",
     "react-query": "^3.39.3",
     "react-redux": "^9.1.2",
+    "redux": "^5.0.1",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^3.1.0",
     "tailwindcss": "^3.4.3",
