+++ ui/src/middleware.ts
@@ -1,0 +1,4 @@
+export { default } from "next-auth/middleware";
+export const config = {
+  matcher: ["/dashboard"],
+};
