+++ ui/src/utils/auth.ts
@@ -20,6 +20,7 @@ export const authOptions: NextAuthOptions = {
 
         // If no error and we have user data, return it
         if (res.ok && user) {
+          console.log(res);
           return user;
         }
         // Return null if user data could not be retrieved
